+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Add
+
+- Background image in layout
+
 ## [0.0.39] - 2024-07-11
 
 - Add remaining tickets on Slack notification